Ensure that fibrous is initialized at most once per process
Currently two packages cannot independently depend on fibrous without causing a
crash. When the second copy of fibrous is initialized, it tries to redefine
`Object/Function::sync/future`, which throws an error since the first copy of
fibrous defined those properties to be non-configurable. But even if they were
to be configurable, the second copy of fibrous would not be able to get a
reference to node-fiber's `Function::future` (because the first copy of fibrous
overrides it).

This diff's approach is similar to node-fiber's guard against double-inclusion,
which uses the `process` object to ensure the module is initialized at most
once.